trying to avoid definition collisions
